% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpm.R
\name{collect_edges.cpm}
\alias{collect_edges.cpm}
\title{Collect selected edges from a CPM fit}
\usage{
\method{collect_edges}{cpm}(x, ...)
}
\arguments{
\item{x}{A \code{cpm} object.}

\item{...}{For future extension. Currently ignored.}
}
\value{
A logical matrix storing the selected positive and negative edges.
}
\description{
Collect selected edges from a CPM fit
}
