% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm-collect.R
\name{collect_edges}
\alias{collect_edges}
\title{Collect stored CPM edge masks}
\usage{
collect_edges(x, ...)
}
\arguments{
\item{x}{An object with stored CPM edges.}

\item{...}{Unused.}
}
\value{
Stored CPM edge masks.
}
\description{
Collect stored CPM edge masks
}
