% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpm-resamples.R
\name{collect_metrics.cpm_resamples}
\alias{collect_metrics.cpm_resamples}
\title{Collect fold-level metrics from CPM resamples}
\usage{
\method{collect_metrics}{cpm_resamples}(x, ...)
}
\arguments{
\item{x}{A \code{cpm_resamples} object.}

\item{...}{For future extension. Currently ignored.}
}
\value{
A \link[tibble:tibble-package]{tibble} with one row per fold.
}
\description{
Collect fold-level metrics from CPM resamples
}
