% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{collect_predictions.cpm_resamples}
\alias{collect_predictions.cpm_resamples}
\title{Collect observation-level predictions from CPM resamples}
\usage{
\method{collect_predictions}{cpm_resamples}(x, ...)
}
\arguments{
\item{x}{A \code{cpm_resamples} object.}

\item{...}{For future extension. Currently ignored.}
}
\value{
A \link[tibble:tibble-package]{tibble} with one row per observation.
}
\description{
Collect observation-level predictions from CPM resamples
}
