% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm-params.R
\name{cpm_bias_correct}
\alias{cpm_bias_correct}
\title{CPM tuning parameter: bias correction}
\usage{
cpm_bias_correct(values = c(TRUE, FALSE))
}
\arguments{
\item{values}{Allowed bias-correction values.}
}
\value{
A qualitative dials parameter object.
}
\description{
CPM tuning parameter: bias correction
}
