% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm-metrics.R
\name{cpm_cor}
\alias{cpm_cor}
\alias{cpm_cor_vec}
\title{CPM Pearson correlation metric}
\usage{
cpm_cor(data, ...)

cpm_cor_vec(truth, estimate, na_rm = TRUE, case_weights = NULL, ...)
}
\arguments{
\item{data}{A data frame containing truth and estimate columns.}

\item{...}{Unused.}

\item{truth}{The column identifier for the true outcome.}

\item{estimate}{The column identifier for the predicted outcome.}

\item{na_rm}{Should missing values be removed before computing the metric?}

\item{case_weights}{Not used.}
}
\value{
A tibble with a single correlation estimate.
}
\description{
CPM Pearson correlation metric
}
