% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpm-reg.R
\name{cpm_engine_fit}
\alias{cpm_engine_fit}
\title{Internal parsnip engine bridge for \code{cpm_reg()}}
\usage{
cpm_engine_fit(
  x,
  y,
  thresh_method = "alpha",
  thresh_level = 0.01,
  bias_correct = TRUE,
  network = "both"
)
}
\arguments{
\item{x}{Predictor matrix.}

\item{y}{Outcome vector.}

\item{thresh_method, thresh_level, bias_correct, network}{CPM engine
parameters.}
}
\value{
A \code{cpm_fit} object.
}
\description{
Internal parsnip engine bridge for \code{cpm_reg()}
}
