% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm-params.R
\name{cpm_network}
\alias{cpm_network}
\title{CPM tuning parameter: network}
\usage{
cpm_network(values = c("both", "pos", "neg"))
}
\arguments{
\item{values}{Allowed network values.}
}
\value{
A qualitative dials parameter object.
}
\description{
CPM tuning parameter: network
}
