% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/objects.R
\name{cpm_resamples}
\alias{cpm_resamples}
\title{cpm_resamples Resampling Object}
\description{
A \code{cpm_resamples} object is returned by \code{\link[=fit_resamples]{fit_resamples()}} and stores
fold-level outputs from resampling.
}
\section{Structure}{

A \code{cpm_resamples} object is a list with the following elements:
\describe{
\item{\code{spec}}{The originating \code{cpm_spec} object.}
\item{\code{folds}}{List of assessment-row indices for each fold.}
\item{\code{metrics}}{Data frame of fold-level performance metrics.}
\item{\code{predictions}}{Data frame of observation-level predictions with
fold IDs.}
\item{\code{edges}}{Stored edge output based on \code{return_edges}
(\code{NULL}/matrix/array).}
\item{\code{params}}{Parameter list used for the resampling run.}
}
}

\seealso{
\code{\link[=fit_resamples]{fit_resamples()}}, \code{\link[=collect_metrics]{collect_metrics()}}, \code{\link[=collect_predictions]{collect_predictions()}},
\code{\link[=collect_edges]{collect_edges()}}, \code{\link[=summary.cpm_resamples]{summary.cpm_resamples()}}
}
