% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm-metrics.R
\name{cpm_spearman}
\alias{cpm_spearman}
\alias{cpm_spearman_vec}
\title{CPM Spearman correlation metric}
\usage{
cpm_spearman(data, ...)

cpm_spearman_vec(truth, estimate, na_rm = TRUE, case_weights = NULL, ...)
}
\arguments{
\item{data}{A data frame containing truth and estimate columns.}

\item{...}{Unused.}

\item{truth}{The column identifier for the true outcome.}

\item{estimate}{The column identifier for the predicted outcome.}

\item{na_rm}{Should missing values be removed before computing the metric?}

\item{case_weights}{Not used.}
}
\value{
A tibble with a single rank-correlation estimate.
}
\description{
CPM Spearman correlation metric
}
