% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tm-params.R
\name{cpm_thresh_method}
\alias{cpm_thresh_method}
\title{CPM tuning parameter: threshold method}
\usage{
cpm_thresh_method(values = c("alpha", "sparsity"))
}
\arguments{
\item{values}{Allowed threshold methods.}
}
\value{
A qualitative dials parameter object.
}
\description{
CPM tuning parameter: threshold method
}
