% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{fit_resamples}
\alias{fit_resamples}
\title{Fit a model specification on resamples}
\usage{
fit_resamples(object, ...)
}
\arguments{
\item{object}{A model specification object.}

\item{...}{Additional arguments passed to method implementations.}
}
\value{
A resampling result object.
}
\description{
Generic for fitting model specifications across resamples.
}
