% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpm.R
\name{tidy.cpm}
\alias{tidy.cpm}
\title{Tidy a \code{cpm} object}
\usage{
\method{tidy}{cpm}(x, ..., component = c("performance", "edges"))
}
\arguments{
\item{x}{A \code{cpm} object.}

\item{...}{Additional arguments passed to \code{summary()}.}

\item{component}{A character vector indicating the component to tidy.}
}
\value{
A \link[tibble:tibble-package]{tibble} with columns storing parameters of
the \code{cpm} object and further columns depending on the \code{component}
argument.
}
\description{
Tidy a \code{cpm} object
}
