% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpm.R
\name{cpm}
\alias{cpm}
\title{Perform Connectome-based Predictive Modeling (CPM)}
\usage{
cpm(
  conmat,
  behav,
  ...,
  thresh_method = c("alpha", "sparsity"),
  thresh_level = 0.01,
  kfolds = NULL,
  bias_correct = TRUE
)
}
\arguments{
\item{conmat}{A matrix of connectome data. Observations in row, edges in
column (assumed that duplicated edges are removed).}

\item{behav}{A numeric vector contains behavior data. Length must equal to
number of observations in \code{conmat}.}

\item{...}{For future extension. Currently ignored.}

\item{thresh_method, thresh_level}{The threshold method and level used in edge
selection. If method is set to be \code{"alpha"}, the edge selection is based on
the critical value of correlation coefficient. If method is set to be
\code{"sparsity"}, the edge selection is based on the quantile of correlation
coefficient, thus network sparsity is controlled.}

\item{kfolds}{Folds number of cross-validation. If \code{NULL}, it will be set to
be equal to the number of observations, i.e., leave-one-subject-out.}

\item{bias_correct}{Logical value indicating if the connectome data should be
bias-corrected. If \code{TRUE}, the connectome data will be centered and scaled
to have unit variance based on the training data.}
}
\value{
An S3 object of class \code{cpm}, which is a list containing the following
components:

\item{real}{A numeric vector of observed behavior data.}

\item{pred}{A matrix of predicted behavior data. Rows are observations and
columns are different models.}

\item{edges_prop}{A list of selection proportion of edges in each network,
i.e., positive and negative networks.}

\item{cor}{A list of correlation test results between predicted behavior
data and observed behavior data.}

The object also contains the following attributes:

\item{thresh_method}{The threshold method used in edge selection.}

\item{thresh_level}{The threshold level used in edge selection.}

\item{kfolds}{The folds number of cross-validation.}

\item{bias_correct}{Logical value indicating if the connectome data should
be bias-corrected.}
}
\description{
The connectome-based predictive modeling (CPM) is a data-driven approach to
predict individual behavior from brain connectivity data. The CPM is based on
the linear regression model, where the brain connectivity data are used as
predictors and the behavior data are used as response.
}
\details{
The CPM is implemented in the following steps:
\enumerate{
\item Select edges based on the correlation between brain connectivity and
behavior data.
\item Train a linear regression model using the selected edges.
\item Predict behavior data using the trained model.
\item Repeat steps 1-3 using different folds of cross-validation.
}
}
\examples{
conmat <- matrix(rnorm(100 * 100), nrow = 100)
behav <- rnorm(100)
cpm(conmat, behav)
# use different threshold method and level
cpm(conmat, behav, thresh_method = "sparsity", thresh_level = 0.05)
# use a 10-fold cross-validation
cpm(conmat, behav, kfolds = 10)
}
