% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core-objects.R
\name{cpm_fit}
\alias{cpm_fit}
\title{CPM engine fit object}
\arguments{
\item{x}{A \code{cpm_fit} object.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Internal engine fit object returned by the \code{cpmr} parsnip engine.
}
