% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit.R
\name{cpm_spec}
\alias{cpm_spec}
\alias{fit.cpm_spec}
\title{Define a CPM model specification}
\usage{
cpm_spec(
  thresh_method = c("alpha", "sparsity"),
  thresh_level = 0.01,
  kfolds = NULL,
  bias_correct = TRUE,
  return_edges = c("sum", "none", "all"),
  na_action = c("fail", "exclude")
)

\method{fit}{cpm_spec}(object, conmat, behav, ..., covariates = NULL, confounds = NULL)
}
\arguments{
\item{thresh_method, thresh_level}{The threshold method and level used in edge
selection. If method is set to be \code{"alpha"}, the edge selection is based on
the critical value of correlation coefficient. If method is set to be
\code{"sparsity"}, the edge selection is based on the quantile of correlation
coefficient, thus network sparsity is controlled.}

\item{kfolds}{Folds number of cross-validation. If \code{NULL}, it will be set to
be equal to the number of observations, i.e., leave-one-subject-out.}

\item{bias_correct}{Logical value indicating if the connectome data should be
bias-corrected. If \code{TRUE}, the connectome data will be centered and scaled
to have unit variance based on the training data before model fitting and
prediction. See Rapuano et al. (2020) for more details.}

\item{return_edges}{A character string indicating the return value of the
selected edges. If \code{"none"}, no edges are returned. If \code{"sum"}, the sum of
selected edges across folds is returned. If \code{"all"}, the selected edges for
each fold is returned, which is a 3D array and memory-consuming.}

\item{na_action}{A character string indicating the action when missing values
are found in \code{behav}. If \code{"fail"}, an error will be thrown. If \code{"exclude"},
missing values will be excluded from the analysis but kept in the output.}

\item{object}{A \code{cpm_spec} object.}

\item{conmat}{A matrix of connectome data. Observations in row, edges in
column (assumed that duplicated edges are removed).}

\item{behav}{A numeric vector contains behavior data. Length must equal to
number of observations in \code{conmat}. Note \code{behav} could also be a row/column
matrix, which will be converted to a vector using \code{\link[=drop]{drop()}}.}

\item{...}{For future extension. Currently ignored.}

\item{covariates}{A matrix of covariates. Observations in row, variables in
column. If \code{NULL}, no covariates are used. Note if a vector is provided, it
will be converted to a column matrix.}

\item{confounds}{Deprecated alias of \code{covariates}.}
}
\value{
A \code{cpm_spec} object storing parameters for later fitting.

A fitted \code{cpm} object.
}
\description{
Create a lightweight specification object that stores the modeling
parameters required to fit a connectome-based predictive model later with
\code{\link[=fit]{fit()}}.
}
\examples{
spec <- cpm_spec(kfolds = 10, return_edges = "sum")
spec

conmat <- matrix(rnorm(100 * 100), nrow = 100)
behav <- rnorm(100)
fit(spec, conmat = conmat, behav = behav)
}
