% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpmr.R
\docType{package}
\name{cpmr-package}
\alias{cpmr}
\alias{cpmr-package}
\title{cpmr--package}
\description{
Implements Connectome-based Predictive Modeling (CPM) for
cross-sectional and time series neuroimaging data in R. Methods include
distance correlation, import tools for CONN connectivity matrices,
and a suite of CPM functions.
}
\author{
Alexander P. Christensen <alexpaulchristensen@gmail.com> & Christopher N. Wahlheim <cnwahlheim@gmail.com>
}
