% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpmr-package.R
\docType{package}
\name{cpmr-package}
\alias{cpmr}
\alias{cpmr-package}
\title{cpmr: Connectome Predictive Modelling in R}
\description{
Connectome Predictive Modelling (CPM) (Shen et al. (2017) \doi{10.1038/nprot.2016.178}) is a method to predict individual differences in behaviour from brain functional connectivity. 'cpmr' provides a simple yet efficient implementation of this method.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/psychelzh/cpmr}
  \item \url{https://psychelzh.github.io/cpmr/}
  \item Report bugs at \url{https://github.com/psychelzh/cpmr/issues}
}

}
\author{
\strong{Maintainer}: Liang Zhang \email{psychelzh@outlook.com} (\href{https://orcid.org/0000-0001-9041-1150}{ORCID})

}
\keyword{internal}
