% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.cpm}
\alias{summary.cpm}
\alias{print.cpm_summary}
\title{Summary of a cpm object.}
\usage{
\method{summary}{cpm}(object, edge_level = 0.5, ...)

\method{print}{cpm_summary}(x, ...)
}
\arguments{
\item{object}{An object of class \code{cpm}.}

\item{edge_level}{A numeric value between 0 and 1 indicating the proportional
threshold for edge selection.}
}
\description{
This function provides a summary of a \code{cpm} object, including the
prediction performance and the selected edges.
}
