% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cpm-resamples.R
\name{summary.cpm_resamples}
\alias{summary.cpm_resamples}
\alias{print.cpm_resamples_summary}
\title{Summary of a \code{cpm_resamples} object}
\usage{
\method{summary}{cpm_resamples}(object, ...)

\method{print}{cpm_resamples_summary}(x, ...)
}
\arguments{
\item{object}{An object of class \code{cpm_resamples}.}

\item{...}{For future extension. Currently ignored.}

\item{x}{An object of class \code{cpm_resamples_summary}.}
}
\value{
A \code{cpm_resamples_summary} object with aggregated fold-level
performance and edge-selection rates.
}
\description{
Summary of a \code{cpm_resamples} object
}
